Further split up auditLogTableConstants
