+++ src/app/admin/auditLogTable/format.ts
@@ -1,0 +1,23 @@
+import { formatBoolean, formatDateTime, formatJson } from "@/common/format";
+import { AuditLogCountError, AuditLogError } from "./fetchAuditLogData";
+
+export const auditLogTableColumnDisplayFunctions = {
+    createdAt: formatDateTime,
+    wasSuccess: formatBoolean,
+    content: formatJson,
+};
+
+export const getAuditLogErrorMessage = (error: AuditLogError | AuditLogCountError): string => {
+    let errorMessage: string = "";
+    switch (error.type) {
+        case "failedAuditLogFetch":
+            errorMessage = "Failed to fetch audit log.";
+            break;
+        case "failedAuditLogCountFetch":
+            errorMessage = "Failed to fetch audit log count.";
+            break;
+        case "nullCount":
+            errorMessage = "Audit log table empty.";
+    }
+    return `${errorMessage} Log ID: ${error.logId}`;
+};
