+++ src/app/admin/auditLogTable/rowsPerPageConstants.ts
@@ -1,0 +1,2 @@
+export const defaultNumberOfAuditLogRowsPerPage = 10;
+export const numberOfAuditLogRowsPerPageOption = [10, 25, 50, 100];
