+++ src/app/admin/auditLogTable/types.ts
@@ -1,0 +1,43 @@
+import { Json } from "@/databaseTypesFile";
+import { Schema } from "@/databaseUtils";
+
+export interface AuditLogRow {
+    action: string;
+    clientId: string;
+    collectionCentreId: string;
+    content: Json;
+    createdAt: string;
+    eventId: string;
+    listHotelId: string;
+    listId: string;
+    logId: string;
+    packingSlotId: string;
+    parcelId: string;
+    profileId: string;
+    statusOrder: string;
+    userId: string;
+    wasSuccess: boolean;
+    websiteData: string;
+}
+
+export const convertAuditLogResponseToAuditLogRow = (
+    auditLogResponse: Schema["audit_log"][]
+): AuditLogRow[] =>
+    auditLogResponse.map((auditLogResponseRow) => ({
+        action: auditLogResponseRow.action ?? "",
+        clientId: auditLogResponseRow.client_id ?? "",
+        collectionCentreId: auditLogResponseRow.collection_centre_id ?? "",
+        content: auditLogResponseRow.content ?? "",
+        createdAt: auditLogResponseRow.created_at ?? "",
+        eventId: auditLogResponseRow.event_id ?? "",
+        listHotelId: auditLogResponseRow.list_hotel_id ?? "",
+        listId: auditLogResponseRow.list_id ?? "",
+        logId: auditLogResponseRow.log_id ?? "",
+        packingSlotId: auditLogResponseRow.packing_slot_id ?? "",
+        parcelId: auditLogResponseRow.parcel_id ?? "",
+        profileId: auditLogResponseRow.profile_id ?? "",
+        statusOrder: auditLogResponseRow.status_order ?? "",
+        userId: auditLogResponseRow.user_id ?? "",
+        wasSuccess: auditLogResponseRow.wasSuccess ?? "",
+        websiteData: auditLogResponseRow.website_data ?? "",
+    }));
